# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kavach Server API",
        "title": "Kavach Server API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
...
        "/organisations/{organisation_id}": {
            "get": {
                "description": "Get organisation by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation"
                ],
                "summary": "Show a organisation by id",
                "operationId": "get-organisation-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organisation.orgWithRole"
                        }
                    }
                }
            },
            "put": {
                "description": "Update organisation by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation"
                ],
                "summary": "Update a organisation by id",
                "operationId": "update-organisation-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Organisation Object",
                        "name": "Organisation",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Organisation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organisation.orgWithRole"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete organisation by ID",
                "tags": [
                    "Organisation"
                ],
                "summary": "Delete a organisation",
                "operationId": "delete-organisation-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/organisations/{organisation_id}/applications/{application_id}/tokens": {
            "get": {
                "description": "Show a application tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationApplicationsTokens"
                ],
                "summary": "List application tokens",
                "operationId": "get-organisation-application-tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/token.paging"
                        }
                    }
                }
            },
            "post": {
                "description": "Create application token by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationApplicationsTokens"
                ],
                "summary": "Show a application token by id",
                "operationId": "create-organisation-application-token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Application Token Body",
                        "name": "ApplicationTokenBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/token.applicationToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApplicationToken"
                        }
                    }
                }
            }
        },
...
        "/organisations/{organisation_id}/policy": {
            "get": {
                "description": "get policy for an organisation using organisation_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationPolicy"
                ],
                "summary": "get policy for an organisation using organisation_id",
                "operationId": "list-organisation-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Policy",
                        "name": "OrganisationRoleBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Policy"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Create policy for an organisation using organisation_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationPolicy"
                ],
                "summary": "Create policy for an organisation using organisation_id",
                "operationId": "create-organisation-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/organisations/{organisation_id}/policy/allowed": {
            "post": {
                "description": "check if a user has permission to perform an action on a resource",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationPolicy"
                ],
                "summary": "check if a user has permission to perform an action on a resource",
                "operationId": "check-organisation-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Application": {
            "type": "object",
            "properties": {
                "application_role": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApplicationRole"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "medium_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "spaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApplicationToken"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.ApplicationRole": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.ApplicationToken": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.Medium": {
            "type": "object",
            "properties": {
                "alt_text": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Organisation": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "organisation_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationUser"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationRole"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
...
        "model.OrganisationUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Policy": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "effect": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "organisation.orgWithRole": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "organisation_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationUser"
                    }
                },
                "permission": {
                    "$ref": "#/definitions/model.OrganisationUser"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationRole"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "token.applicationToken": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "token.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApplicationToken"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}